implement new history view
